Modificación de los servicios de movimientos varios asociados y aplicados de la ficha de certificación

+++ siga-core/src/main/java/org/itcgae/siga/DTOs/scs/AsuntoPorMovimientoItem.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class AsuntoPorMovimientoItem {
 
     private String asunto;
-    private Short idMovimiento;
+    private Long idMovimiento;
 }
